+++ test/integ/cli-full-lifecycle/scaleway.sh
@@ -27,10 +27,19 @@ function create_destroy_scaleway() {
     $cloudypad_cmd list | grep $instance_name
 
     check_instance_status $instance_name
+    
+    $cloudypad_cmd get $instance_name
 
+    # Stop/start twice to check idempotency
+    $cloudypad_cmd stop $instance_name --wait
     $cloudypad_cmd stop $instance_name --wait
 
+    $cloudypad_cmd get $instance_name
+
     $cloudypad_cmd start $instance_name --wait
+    $cloudypad_cmd start $instance_name --wait
+
+    $cloudypad_cmd get $instance_name
 
     $cloudypad_cmd restart $instance_name --wait