+++ test/unit/core/state/writer.spec.ts
@@ -10,44 +10,45 @@ import { AwsInstanceStateV1, AwsStateParser } from '../../../../src/providers/aw
 import lodash from 'lodash'
 import { LocalStateSideEffect } from '../../../../src/core/state/side-effects/local'
 import { DUMMY_V1_ROOT_DATA_DIR } from '../../utils'
+import { InstanceEventEnum } from '../../../../src/core/state/state'
+import { AnonymousStateParser, GenericStateParser } from '../../../../src/core/state/parser'
+import { DummyInstanceStateV1, DummyStateParser } from '../../../../src/providers/dummy/state'
 
 describe('StateWriter', function () {
 
-    const instanceName = "aws-dummy"
+    const testInstanceName = "dummy-provider-state"
 
     // create a test writer using a temp directory as data dir
-    async function getTestWriter(): Promise<{ dataDir: string, writer: StateWriter<AwsInstanceStateV1> }> {
+    async function getTestWriter(): Promise<{ dataDir: string, writer: StateWriter<DummyInstanceStateV1> }> {
         const dataDir = mkdtempSync(path.join(tmpdir(), 'statewriter-test-'))
 
         // load a dummy state and copy it into our test writer
         const loader = new StateLoader({ 
             sideEffect: new LocalStateSideEffect({ dataRootDir: DUMMY_V1_ROOT_DATA_DIR})
         })
-        const state = await loader.loadInstanceState(instanceName)
-        const awState = new AwsStateParser().parse(state)
+        const state = await loader.loadInstanceState(testInstanceName)
+        const dummyState = new DummyStateParser().parse(state)
 
         // create a test writer and persist the state
-        const writer = new StateWriter<AwsInstanceStateV1>({
-            sideEffect: new LocalStateSideEffect({ dataRootDir: dataDir })
+        const writer = new StateWriter<DummyInstanceStateV1>({
+            sideEffect: new LocalStateSideEffect({ dataRootDir: dataDir }),
+            stateParser: new DummyStateParser()
         })
-        writer.setState(awState)
-        await writer.persistStateNow()
+        await writer.setState(dummyState)
 
         return { dataDir: dataDir, writer: writer }
     }
 
     // Load state from given data dir to compare with expected result
     function loadResultPersistedState(dataDir: string){
-        const filePath = path.resolve(path.join(dataDir, "instances", instanceName, "state.yml"))
+        const filePath = path.resolve(path.join(dataDir, "instances", testInstanceName, "state.yml"))
         return yaml.parse(fs.readFileSync(filePath, 'utf-8'))
     }
 
     it('should write on disk state held in memory', async function () {
         const { dataDir, writer } = await getTestWriter()
 
-        await writer.persistStateNow()
-
-        const expected = writer.cloneState()
+        const expected = await writer.getCurrentState(testInstanceName)
         const result = loadResultPersistedState(dataDir)
         assert.deepStrictEqual(expected, result)
     })
@@ -55,12 +56,13 @@ describe('StateWriter', function () {
     it('should update provision input', async function () {
         const { dataDir, writer } = await getTestWriter()
 
-        await writer.updateProvisionInput({ 
+        await writer.updateProvisionInput(testInstanceName, { 
             diskSize: 999,
         })
 
+        const clonedState = await writer.getCurrentState(testInstanceName)
         const expected = lodash.merge(
-            writer.cloneState(),
+            clonedState,
             {
                 provision: {
                     input: {
@@ -77,12 +79,13 @@ describe('StateWriter', function () {
     it('should update configuration input', async function () {
         const { dataDir, writer } = await getTestWriter()
 
-        await writer.updateConfigurationInput({ 
+        await writer.updateConfigurationInput(testInstanceName, { 
             dummyConfig: "bar",
         })
 
+        const clonedState = await writer.getCurrentState(testInstanceName)
         const expected = lodash.merge(
-            writer.cloneState(),
+            clonedState,
             {
                 configuration: {
                     input: {
@@ -99,15 +102,18 @@ describe('StateWriter', function () {
     it('should set provision input', async function () {
         const { dataDir, writer } = await getTestWriter()
 
+        const currentState = await writer.getCurrentState(testInstanceName)
+
         const newProvInput = { 
-            ...writer.cloneState().provision.input,
+            ...currentState.provision.input,
             diskSize: 1234,
             instanceType: "g5.xlarge"
         }
-        await writer.setProvisionInput(newProvInput)
+        await writer.setProvisionInput(testInstanceName, newProvInput)
 
+        const stateAfterUpdate = await writer.getCurrentState(testInstanceName)
         const expected = lodash.merge(
-            writer.cloneState(),
+            stateAfterUpdate,
             {
                 provision: {
                     input: newProvInput
@@ -122,14 +128,17 @@ describe('StateWriter', function () {
     it('should set configuration input', async function () {
         const { dataDir, writer } = await getTestWriter()
 
+        const currentState = await writer.getCurrentState(testInstanceName)
+
         const newConfInput = { 
-            ...writer.cloneState().configuration.input,
+            ...currentState.configuration.input,
             dummyConf: "foo",
         }
-        await writer.setConfigurationInput(newConfInput)
+        await writer.setConfigurationInput(testInstanceName, newConfInput)
 
+        const stateAfterUpdate = await writer.getCurrentState(testInstanceName)
         const expected = lodash.merge(
-            writer.cloneState(),
+            stateAfterUpdate,
             {
                 configuration: {
                     input: newConfInput
@@ -144,17 +153,19 @@ describe('StateWriter', function () {
     it('should set configuration output', async function () {
         const { dataDir, writer } = await getTestWriter()
 
-        const output = {
-            dataDiskConfigured: true
+        const expectedOutput: DummyInstanceStateV1['configuration']['output'] = {
+            configuredAt: new Date("2025-01-01T01:00:00Z").getTime(),
+            dataDiskConfigured: true,
         }
 
-        await writer.setConfigurationOutput(output)
+        await writer.setConfigurationOutput(testInstanceName, expectedOutput)
 
+        const clonedState = await writer.getCurrentState(testInstanceName)
         const expected = lodash.merge(
-            writer.cloneState(),
+            clonedState,
             {
                 configuration: {
-                    output: output
+                    output: expectedOutput
                 }
             }
         )
@@ -166,18 +177,21 @@ describe('StateWriter', function () {
     it('should set provision output', async function () {
         const { dataDir, writer } = await getTestWriter()
 
-        const output = {
+        const expectedOutput: DummyInstanceStateV1['provision']['output'] = {
             host: "1.2.3.4",
-            instanceId: "i-123456758"
+            instanceId: "i-123456758",
+            provisionedAt: new Date("2025-01-01T01:00:00Z").getTime(),
+            dataDiskId: "ssd-123456758",
         }
 
-        await writer.setProvisionOutput(output)
+        await writer.setProvisionOutput(testInstanceName, expectedOutput)
 
+        const clonedState = await writer.getCurrentState(testInstanceName)
         const expected = lodash.merge(
-            writer.cloneState(),
+            clonedState,
             {
                 provision: {
-                    output: output
+                    output: expectedOutput
                 }
             }
         )
@@ -190,13 +204,13 @@ describe('StateWriter', function () {
         const { dataDir, writer } = await getTestWriter()
 
         // check if state file exists
-        const stateDirPath = path.resolve(path.join(dataDir, "instances", instanceName))
+        const stateDirPath = path.resolve(path.join(dataDir, "instances", testInstanceName))
         const stateFilePath = path.resolve(path.join(stateDirPath, "state.yml"))
         assert.ok(fs.existsSync(stateFilePath))
         assert.ok(fs.existsSync(stateDirPath))
 
         // Call the destroyState method
-        await writer.destroyState()
+        await writer.destroyState(testInstanceName)
 
         // Check state file and parent dir no longer exists
         const fileExists = fs.existsSync(stateFilePath)
@@ -205,6 +219,101 @@ describe('StateWriter', function () {
         const parentDirExists = fs.existsSync(stateDirPath)
         assert.strictEqual(parentDirExists, false)
     })
-    
-})
 
+    it('should add event to state (up to max events)', async function () {
+        const { dataDir, writer } = await getTestWriter()
+        const eventDate = new Date("2025-01-01T01:00:00Z")
+
+        // try to load without events
+        const rawStateNoEvent = await loadResultPersistedState(dataDir)
+        const stateParser = new AnonymousStateParser()
+        const stateNoEvent = stateParser.parse(rawStateNoEvent)
+        assert.ok(stateNoEvent.events === undefined)
+
+        // add a single event
+        await writer.addEvent(testInstanceName, InstanceEventEnum.ProvisionBegin, eventDate)
+
+        const rawStateOneEvent = await loadResultPersistedState(dataDir)
+        const stateOneEvent = stateParser.parse(rawStateOneEvent)
+        assert.strictEqual(stateOneEvent.events?.length, 1)
+        assert.strictEqual(stateOneEvent.events?.[0].type, InstanceEventEnum.ProvisionBegin)
+
+        // add 9 more events (max 10)
+        await writer.addEvent(testInstanceName, InstanceEventEnum.ProvisionEnd,       new Date(eventDate.getTime() + 1))
+        await writer.addEvent(testInstanceName, InstanceEventEnum.ConfigurationBegin, new Date(eventDate.getTime() + 2))
+        await writer.addEvent(testInstanceName, InstanceEventEnum.ConfigurationEnd,   new Date(eventDate.getTime() + 3))
+        await writer.addEvent(testInstanceName, InstanceEventEnum.StartBegin,         new Date(eventDate.getTime() + 4))
+        await writer.addEvent(testInstanceName, InstanceEventEnum.StartEnd,           new Date(eventDate.getTime() + 5))
+        await writer.addEvent(testInstanceName, InstanceEventEnum.StopBegin,          new Date(eventDate.getTime() + 6))
+        await writer.addEvent(testInstanceName, InstanceEventEnum.StopEnd,            new Date(eventDate.getTime() + 7))
+        await writer.addEvent(testInstanceName, InstanceEventEnum.DestroyBegin,       new Date(eventDate.getTime() + 8))
+        await writer.addEvent(testInstanceName, InstanceEventEnum.DestroyEnd,         new Date(eventDate.getTime() + 9))
+
+        const rawStateTenEvents = await loadResultPersistedState(dataDir)
+        const stateTenEvents = stateParser.parse(rawStateTenEvents)
+        assert.ok(stateTenEvents.events)
+        assert.strictEqual(stateTenEvents.events?.length, 10)
+
+        const tenEvents = lodash.cloneDeep(stateTenEvents.events)
+        tenEvents.sort((a, b) => a.timestamp - b.timestamp)
+        assert.strictEqual(tenEvents[0].type, InstanceEventEnum.ProvisionBegin)
+        assert.strictEqual(tenEvents[0].timestamp, eventDate.getTime())
+        assert.strictEqual(tenEvents[1].type, InstanceEventEnum.ProvisionEnd)
+        assert.strictEqual(tenEvents[1].timestamp, eventDate.getTime() + 1)
+        assert.strictEqual(tenEvents[2].type, InstanceEventEnum.ConfigurationBegin)
+        assert.strictEqual(tenEvents[2].timestamp, eventDate.getTime() + 2)
+        assert.strictEqual(tenEvents[3].type, InstanceEventEnum.ConfigurationEnd)
+        assert.strictEqual(tenEvents[3].timestamp, eventDate.getTime() + 3)
+        assert.strictEqual(tenEvents[4].type, InstanceEventEnum.StartBegin)
+        assert.strictEqual(tenEvents[4].timestamp, eventDate.getTime() + 4)
+        assert.strictEqual(tenEvents[5].type, InstanceEventEnum.StartEnd)
+        assert.strictEqual(tenEvents[5].timestamp, eventDate.getTime() + 5)
+        assert.strictEqual(tenEvents[6].type, InstanceEventEnum.StopBegin)
+        assert.strictEqual(tenEvents[6].timestamp, eventDate.getTime() + 6)
+        assert.strictEqual(tenEvents[7].type, InstanceEventEnum.StopEnd)
+        assert.strictEqual(tenEvents[7].timestamp, eventDate.getTime() + 7)
+        assert.strictEqual(tenEvents[8].type, InstanceEventEnum.DestroyBegin)
+        assert.strictEqual(tenEvents[8].timestamp, eventDate.getTime() + 8)
+        assert.strictEqual(tenEvents[9].type, InstanceEventEnum.DestroyEnd)
+        assert.strictEqual(tenEvents[9].timestamp, eventDate.getTime() + 9)
+
+        // add 11th event, should remove oldest event
+        await writer.addEvent(testInstanceName, InstanceEventEnum.ProvisionBegin, new Date(eventDate.getTime() + 10))
+        const rawStateElevenEvents = await loadResultPersistedState(dataDir)
+        const stateElevenEvents = stateParser.parse(rawStateElevenEvents)
+
+        assert.ok(stateElevenEvents.events)
+        assert.strictEqual(stateElevenEvents.events?.length, 10)
+
+        const elevenEvents = lodash.cloneDeep(stateElevenEvents.events)
+        elevenEvents.sort((a, b) => a.timestamp - b.timestamp)
+
+        // oldest event should be removed
+        assert.strictEqual(elevenEvents[0].type, InstanceEventEnum.ProvisionEnd)
+        assert.strictEqual(elevenEvents[0].timestamp, eventDate.getTime() + 1)
+
+        assert.strictEqual(elevenEvents[1].type, InstanceEventEnum.ConfigurationBegin)
+        assert.strictEqual(elevenEvents[1].timestamp, eventDate.getTime() + 2)
+
+        assert.strictEqual(elevenEvents[9].type, InstanceEventEnum.ProvisionBegin)
+        assert.strictEqual(elevenEvents[9].timestamp, eventDate.getTime() + 10)
+
+        // again add 11th event, should remove oldest event
+        await writer.addEvent(testInstanceName, InstanceEventEnum.ProvisionEnd, new Date(eventDate.getTime() + 11))
+        const rawStateTwelveEvents = await loadResultPersistedState(dataDir)
+        const stateTwelveEvents = stateParser.parse(rawStateTwelveEvents)
+        assert.ok(stateTwelveEvents.events)
+        assert.strictEqual(stateTwelveEvents.events?.length, 10)
+
+        const twelveEvents = lodash.cloneDeep(stateTwelveEvents.events)
+        twelveEvents.sort((a, b) => a.timestamp - b.timestamp)
+        assert.strictEqual(twelveEvents[0].type, InstanceEventEnum.ConfigurationBegin)
+        assert.strictEqual(twelveEvents[0].timestamp, eventDate.getTime() + 2)
+
+        assert.strictEqual(twelveEvents[1].type, InstanceEventEnum.ConfigurationEnd)
+        assert.strictEqual(twelveEvents[1].timestamp, eventDate.getTime() + 3)
+
+        assert.strictEqual(twelveEvents[9].type, InstanceEventEnum.ProvisionEnd)
+        assert.strictEqual(twelveEvents[9].timestamp, eventDate.getTime() + 11)
+    })
+})