+++ test/unit/utils.ts
@@ -13,6 +13,13 @@ import { STREAMING_SERVER_SUNSHINE } from '../../src/cli/prompter';
 import { CreateCliArgs } from "../../src/cli/command";
 import { ScalewayPulumiOutput } from "../../src/providers/scaleway/pulumi";
 import { CloudypadClient } from "../../src";
+import { AwsInstanceStateV1 } from "../../src/providers/aws/state";
+import { PartialDeep } from "type-fest";
+import * as lodash from "lodash"
+import { CoreConfig } from "../../src/core/config/interface";
+import { DummyProviderClient } from "../../src/providers/dummy/provider";
+import { DummyInstanceStateV1, DummyProvisionInputV1, DummyConfigurationOutputV1 } from "../../src/providers/dummy/state";
+
 /**
  * CommonInstanceInput with as most fields filled as possible while keeping it valid:
  * - privateKeyPath is set but not privateKeyContent
@@ -45,6 +52,9 @@ export const DEFAULT_COMMON_INPUT: CommonInstanceInput = {
             variant: "azerty",
             model: "pc105",
             options: "ctrl:swap_lalt_lctl"
+        },
+        ansible: {
+            additionalArgs: "-t dummytag"
         }
     },
 }
@@ -67,6 +77,7 @@ export const DEFAULT_COMMON_CLI_ARGS: CreateCliArgs = {
     keyboardVariant: DEFAULT_COMMON_INPUT.configuration.keyboard?.variant,
     keyboardModel: DEFAULT_COMMON_INPUT.configuration.keyboard?.model,
     keyboardOptions: DEFAULT_COMMON_INPUT.configuration.keyboard?.options,
+    ansibleAdditionalArgs: DEFAULT_COMMON_INPUT.configuration.ansible?.additionalArgs,
 }
 
 export const DUMMY_SSH_KEY_PATH = path.resolve(__dirname, '..', 'resources', 'ssh-key')
@@ -92,10 +103,12 @@ export const DUMMY_GCP_PULUMI_OUTPUT: GcpPulumiOutput = { instanceName: "dummy-g
  * Dummy output returned by Pulumi during unit test for Scaleway
  */
 export const DUMMY_SCALEWAY_PULUMI_OUTPUT: ScalewayPulumiOutput = { 
-    instanceName: "dummy-scw", 
+    instanceServerName: "dummy-scw", 
     publicIp: "127.0.0.1", 
     instanceServerId: "server-123456789",
-    rootDiskId: "disk-123456789"
+    rootDiskId: "disk-123456789",
+    dataDiskId: null,
+    instanceServerUrn: "urn:xxx"
 }
 
 export const DUMMY_V1_ROOT_DATA_DIR = path.resolve(__dirname, "..", "resources", "states", "v1-root-data-dir")
@@ -135,12 +148,86 @@ const TEST_DATA_ROOT_DIR = createTempTestDir("data-root")
  */
 export function getUnitTestCoreClient(): CloudypadClient{
     return new CloudypadClient({
-        config: {
-            stateBackend: {
-                local: {
-                    dataRootDir: TEST_DATA_ROOT_DIR
-                }
+        config: getUnitTestCoreConfig()
+    })
+}
+
+export function getUnitTestCoreConfig(): CoreConfig {
+    return {
+        stateBackend: {
+            local: { dataRootDir: TEST_DATA_ROOT_DIR }
+        }
+    }
+}
+
+export function getUnitTestDummyProviderClient(): DummyProviderClient {
+    return new DummyProviderClient({
+        config: getUnitTestCoreConfig()
+    })
+}
+
+/**
+ * Create a dummy AWS state that can be used for testing
+ * @param instanceName 
+ * @returns 
+ */
+export function createDummyAwsState(override: PartialDeep<AwsInstanceStateV1>): AwsInstanceStateV1 {
+
+    // clone deep to avoid later operation returned state
+    // to alter DEFAULT_COMMON_INPUT used in this state
+    const awsState: AwsInstanceStateV1 = lodash.cloneDeep({
+        version: "1",
+        provision: {
+            provider: "aws",
+            input: {
+                ...DEFAULT_COMMON_INPUT.provision,
+                diskSize: 100,
+                instanceType: "g4dn.xlarge",
+                region: "eu-west-1",
+                publicIpType: "static",
+                useSpot: true,
             }
+        },
+        name: `dummy-aws-instance-${Date.now()}`,
+        configuration: {
+            configurator: "ansible",
+            input: {
+                ...DEFAULT_COMMON_INPUT.configuration,
+            },
         }
     })
+
+    return lodash.merge(awsState, override)
+}
+
+/**
+ * Create a dummy state that can be used for testing
+ * @param override 
+ * @returns 
+ */
+export function createDummyState(override: PartialDeep<DummyInstanceStateV1>): DummyInstanceStateV1 {
+
+    // clone deep to avoid later operation returned state
+    // to alter DEFAULT_COMMON_INPUT used in this state
+    const dummyState: DummyInstanceStateV1 = lodash.cloneDeep({
+        version: "1",
+        provision: {
+            provider: "dummy",
+            input: {
+                ...DEFAULT_COMMON_INPUT.provision,
+                instanceType: "t2.micro",
+                startDelaySeconds: 5,
+                stopDelaySeconds: 5,
+            }
+        },
+        name: `dummy-instance-${Date.now()}`,
+        configuration: {
+            configurator: "ansible",
+            input: {
+                ...DEFAULT_COMMON_INPUT.configuration,
+            },
+        }
+    });
+
+    return lodash.merge(dummyState, override);
 }